// src/app/api/swagger/models/api-full-deal-dto.ts
/* tslint:disable */
/* eslint-disable */
import { ApiAdditionalAgreementDto } from '../models/api-additional-agreement-dto';
import { ApiApplicationStateEnum } from '../models/api-application-state-enum';
import { ApiContactInfoDto } from '../models/api-contact-info-dto';
import { ApiDealDocumentDto } from '../models/api-deal-document-dto';
import { ApiDealEisIntegrationDataAltDto } from '../models/api-deal-eis-integration-data';
import { ApiDealFormEnum } from '../models/api-deal-form-enum';
import { ApiDealObjectEnum } from '../models/api-deal-object-enum';
import { ApiDealPositionDto } from '../models/api-deal-position-dto';
import { ApiDealStateEnum } from '../models/api-deal-state-enum';
import { ApiOrganizationCountryAltEnum } from '../models/api-organization-country';
import { ApiPartisipantInfoForDealOutSideEShopAltDto } from '../models/api-partisipant-info-for-deal-out-side-e-shop';
export interface ApiFullDealDto {
  AdditionalAgreements?: Array<ApiAdditionalAgreementDto> | null;
  AllowPaperDealForm?: boolean;
  ApplicationId?: number | null;
  ApplicationState?: ApiApplicationStateEnum | null;
  BudgetId?: number | null;
  BuyerAgreementComment?: string | null;
  CancellationReason?: string | null;
  ChangeRequestReason?: string | null;
  ChatId?: number;
  ConclusionDate?: string | null;
  ContactInfo?: ApiContactInfoDto | null;
  ContractExecutionDate?: string | null;
  CustomContractNumber?: string | null;
  CustomerCancellationReason?: string | null;
  DealComment?: string | null;
  DealCustomNumber?: string | null;
  DealEisCustomNumber?: string | null;
  DealObject?: ApiDealObjectEnum;
  DealObjectName?: string | null;
  DealOutSideEshop?: boolean;
  DealPositions?: Array<ApiDealPositionDto> | null;
  DifferenceProtocolReason?: string | null;
  Documents?: Array<ApiDealDocumentDto> | null;
  EisIntegrationData?: ApiDealEisIntegrationDataAltDto | null;
  ExecutionFrom?: string | null;
  ExecutionTo?: string | null;
  ExternalSystemId?: number | null;
  Form?: ApiDealFormEnum;
  Id?: number;
  IsAttachPaperDealDocumentsAllowed?: boolean;
  IsContractSignedOutOfEp?: boolean;
  IsCreateAdditionalAgreementsAllowed?: boolean;
  IsDealExecutionOrTerminationAllowed?: boolean;
  IsPriceChanged?: boolean;
  IsSendContractAllowed?: boolean;
  IsSendDealToEisAllowed?: boolean;
  IsUnitBidding?: boolean;
  ParticipantExternalOrganizationId?: string | null;
  ParticipantGuid?: string;
  ParticipantId?: string | null;
  ParticipantInn?: string | null;
  ParticipantKpp?: string | null;
  ParticipantName?: string | null;
  ParticipantOrganizationCountry?: ApiOrganizationCountryAltEnum;
  ParticipantPhone?: string | null;
  ParticipantSource?: number | null;
  ParticipantWarningMessage?: string | null;
  PartisipantInfo?: ApiPartisipantInfoForDealOutSideEShopAltDto | null;
  Price?: number | null;
  PriceWarningMessage?: string | null;
  ReadOnly?: boolean;
  RefusedByCustomerDate?: string | null;
  RequestReason?: boolean;
  SendToProviderDate?: string | null;
  ShowDownloadFileLinks?: boolean;
  SignByProviderDate?: string | null;
  SignedByOrganizer?: boolean;
  State?: ApiDealStateEnum;
  TradeId?: number;
  TradeIdentificationCode?: string | null;
  TradeLotId?: number;
  WorkGroupIds?: Array<number> | null;
}